+++ remix/app/utils/electronBridge.ts
@@ -198,6 +198,7 @@ export type ThingtimeNodeDeviceRequest =
 				| 'application.hide'
 				| 'application.unhide'
 				| 'system.lock'
+				| 'system.sleep'
 				| 'system.wifi.connect'
 				| 'system.wifi.disconnect'
 				| 'system.wifi.power.set';